adds redis backend support #20
