+++ vishnu/backend/__init__.py
@@ -7,6 +7,27 @@ class BackendType(object):
     GoogleAppEngineNDB = "gae-ndb"
     PythonMemcached = "python-memcached"
     PyMemcache = "pymemcache"
+    Redis = "redis"
+
+
+class VishnuSession(object):
+
+    def __init__(self, expires, last_accessed, data):
+        self._expires = expires
+        self._last_accessed = last_accessed
+        self._data = data
+
+    @property
+    def expires(self):
+        return self._expires
+
+    @property
+    def last_accessed(self):
+        return self._last_accessed
+
+    @property
+    def data(self):
+        return self._data
 
 
 class Base(object):
