+++ vishnu/backend/redis_redis.py
@@ -2,21 +2,22 @@ from vishnu.backend import Base
 from vishnu.backend import VishnuSession
 
 import pickle
-from pymemcache.client.base import Client
+import redis
 
 
 class Backend(Base):
 
     def __init__(self, sid, host, port):
         super(Backend, self).__init__(sid)
-        self._memcache = Client((host, int(port)))
+
+        self._redis = redis.StrictRedis(host=host, port=port, db=0)
 
         self._record = None
 
     def load(self):
 
         if not self._loaded:
-            found_in_cache = self._memcache.get(self._sid)
+            found_in_cache = self._redis.get(self._sid)
 
             if found_in_cache is None:
                 return False
@@ -33,7 +34,7 @@ class Backend(Base):
     def clear(self):
         super(Backend, self).clear()
         if self._sid:
-            self._memcache.delete(self._sid)
+            self._redis.delete(self._sid)
 
     def save(self, sync_only=False):
         import logging
@@ -47,4 +48,4 @@ class Backend(Base):
             self._data
         )
 
-        self._memcache.set(self._sid, pickle.dumps(self._record))
+        self._redis.set(self._sid, pickle.dumps(self._record))
